verifyflow/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==4.1.1
python-multipart==0.0.9
openai==1.51.0
chromadb==0.5.5
pdfplumber==0.11.4
pandas==2.2.2
python-docx==1.1.2
pydantic-settings==2.5.2
python-dotenv==1.0.1